Version 1.4: 1) Split common code of cons and store to agla folder. 2) Edit client types. 3) Add options to menu by permissions. 4) Dynamic client types - for managers. 5) POS - start. 6) More...
Version 1.3:
1) Im haadama closure - enable delete price list.

Version 1.2:
1) Silence php warnings.
2) Change cash delta limit to 400.
3) Bundles comeback!
4) Fresh Siton Catalog.
5) Worker discount.
6) Fix role requirements for internal functions.
7) Fix route to be base on order and not client.
8) Update pricelist from xls - for chef site.

Version 1.1:
1) Coder: use config file for each object.
2) Delivery: select prod from list, change order of lines. Add Collection.
3) Mission: client now select day. Drivers get page per mission.
4) Postcode: get postcode from Israel Post service.
5) Sync zone between sites.
6) Get-total-orders: create single supplier supplies. Send operation to post file.
7) Supplies: finish implementation for units.
8) Tasklist: prime version.
9)

Version 1.0.5:
1) Save payment receipt with each delivery
2) Require login.
3) Count number of lines in order.
4) St

Version 1.0.4
1) Save payment receipt number in delivery.
2) Authrize use of scripts.
3) Get invoice user by email.
4) Set category from search tool.
5) Stop setting supplier-category to products
6) Count line number in delivery.
7) Select term -> auto complete instead of selection.
8) Add needed products cache. Not active for now.

Version 1.0.3
1) Makolet report.
2) Rename business to post.php
3) Add error info when connecting to invoice4u.
4) Bug fix - calculate more the one basket in the delivery.
5) Vat - default yes.
6)

Version 1.0.2
1) Add category to pricelist.
2) Escape ' in product name
3) Fix a bug - qunatity discount line wasn't saved.
4) Get the fresh category from wp_option table.
5) delivery/test.php - to test fresh category.
6) Get the user/password of the database from wp-config.

+++ tools/delivery/get-delivery.php
@@ -1,5 +1,4 @@
 <?php
-require_once '../r-shop_manager.php';
 require_once 'delivery.php';
 require_once '../orders/orders-common.php';
 
@@ -11,69 +10,16 @@ $margin = isset( $_GET["margin"] );
 $d = new Delivery( $id );
 
 $order_id = get_order_id( $id );
+//print current_user_can("edit_shop_orders") ."<br/>";
+//print order_get_customer_id($order_id)."<br/>";
+//print get_current_user_id()."<br/>";
+if ( ( ! current_user_can( "edit_shop_orders" ) ) and ( order_get_customer_id( $order_id ) != get_current_user_id() ) ) {
+	print "אין הרשאה";
+	die( 0 );
+}
 
 print order_info_data( $order_id );
-
 print $d->delivery_text( ImDocumentType::delivery, ImDocumentOperation::show, $margin );
-//$sql = 'select product_name, quantity, vat,price, line_price '
-//        . ' from im_delivery_lines where delivery_id = ' . $id ;
-//
-//
-//$fields = mysql_num_fields ( $export );
-//
-//$data = "<table id=\"del_table\" border=\"1\"><tr><td>פריט</td><td>כמות</td><td>מעם</td><td>מחיר</td><td>סהכ</td></tr>";
-//$total = 0;
-//$vat_total = 0;
-//$line_number = 0;
-//$calc_total = 0;
-//
-//while( $row = mysqli_fetch_row( $result ) )
-//{
-//	$line_number = $line_number + 1;
-//	$line = "<tr>";
-//    	$product_name = $row[0];
-//	$quantity = $row[1];
-//	$vat_line = $row[2];
-//	$item_price = $row[3];
-//	$total_line = $row[4];
-//
-//	// Display item name
-//	$line .= "<td>" . $product_name . '</td>';
-//	$line .= "<td>" . $quantity . "</td>";
-//	$line .= "<td>" . $vat_line . "</td>";
-//	$line .= "<td>" . $item_price . "</td>";
-//	$line .= "<td>" . $total_line . "</td>"; $calc_total += $total_line;
-//
-//	$line .= "</tr>";
-//
-//	$data .= "<tr> ". trim( $line ) . "</tr>";
-//}
-//
-//$sql = 'select vat, total '
-//        . ' from im_delivery where id = ' . $id ;
-//
-//$row = mysqli_fetch_row( $result );
-//
-//$vat_total = $row[0];
-//$total = $row[1];
-//
-//$data .= "<tr><td>עיגול</td><td></td><td></td><td></td><td>" . round(100 * ($total - $calc_total)) /100 . "</td></tr>";
-//$data .= "<tr><td>סהכ ללא מעם</td><td></td><td></td><td></td><td>" . ($total - $vat_total) . "</td></tr>";
-//$data .= "<tr><td>סהכ מעם</td><td></td><td></td><td></td><td>" . $vat_total . "</td></tr>";
-//$data .= "<tr><td>סהכ לתשלום</td><td></td><td></td><td></td><td id=\"total\">" . $total . "</td></tr>";
-//
-//$data = str_replace( "\r" , "" , $data );
-//
-//if ( $data == "" )
-//{
-//    $data = "\n(0) Records Found!\n";
-//}
-//
-//$data .= "</table>";
-//
-//print "$data";
-//
-//print "</form>";
 
 if ( ! $send ) {
 	print '<button id="btn_del" onclick="deleteDelivery()">מחק תעודה</button>';
